feat: restructure api folders and add dispatch middleware
