+++ api/src/lib/octokit/fetch-instrument-data.ts
@@ -1,0 +1,8 @@
+import { octokit, repositoryConfig } from './config.ts';
+
+export const fetchInstrumentData = (path: string) => {
+  return octokit.rest.repos.getContent({
+    path,
+    ...repositoryConfig
+  });
+};
